LoadDefaultData

+++ CalcWin/BusinessLogic/ControllersLogic/SettingsLogic.cs
@@ -1,19 +1,10 @@
-﻿using System;
-using System.Linq;
-using CalcWin.Data;
-using Calculator.Models;
-using CalcWin.Views.Calculator;
-using System.Collections.Generic;
-using CalcWin.BusinessLogic.ControllersValidations;
-using System.Security.Claims;
-using Calculator.BussinesLogic;
-using CalcWin.Models.SettingsViewModels;
+﻿using CalcWin.Data;
 using CalcWin.Data.DefaultData;
-using System.Xml;
-using System.Xml.Serialization;
-using CalcWin.Models;
 using Microsoft.AspNetCore.Http;
 using System.IO;
+using System;
+using System.Collections.Generic;
+using Calculator.Models;
 
 namespace CalcWin.BusinessLogic.ControllersLogic
 {
@@ -28,19 +19,45 @@ namespace CalcWin.BusinessLogic.ControllersLogic
 
         public void LoadDefaultData(IFormFile file)
         {
-
             var path = Path.Combine(
                         Directory.GetCurrentDirectory(), "wwwroot",
                         file.FileName);
 
-            File.Delete(path);
+            if (File.Exists(path))
+            {
+                File.Delete(path);
+            }
 
             using (var stream = new FileStream(path, FileMode.Create))
             {
                 file.CopyToAsync(stream);
             }
 
-            var xml = GenerateDefaultData.LoadXml<DataFile>(path);
+            var defaultData = GenerateDefaultData.LoadXml<DefaultData>(path);
+            File.Delete(path);
+
+            SaveDefaultData(defaultData);
+        }
+
+        private void SaveDefaultData(DefaultData defaultData)
+        {
+            foreach (var fruit in defaultData.Fruits)
+            {
+                db.Fruits.Add(fruit);
+            }
+
+            foreach (var flavor in defaultData.Flavors)
+            {
+                db.Flavors.Add(flavor);
+            }
+
+            db.Supplement.Add(defaultData.Supplements.Water);
+            db.Supplement.Add(defaultData.Supplements.Sugar);
+            db.Supplement.Add(defaultData.Supplements.Acid);
+            db.Supplement.Add(defaultData.Supplements.Yeast);
+            db.Supplement.Add(defaultData.Supplements.YeastFood);
+
+            db.SaveChanges();
         }
     }
 }
